requirements-pipeline.txt
# Pipeline Medallion (rodar com py -3.12)
pandas>=2.0
pyarrow>=14.0
boto3>=1.28
psycopg2-binary>=2.9
